+++ examples/resources/zohomail_domain/resource.tf
@@ -1,0 +1,3 @@
+resource "zohomail_domain" "example" {
+  domain_name = "example.com"
+}